+++ CHANGELOG.md
@@ -19,6 +19,8 @@ For a full diff see [`0.11.0...main`][0.11.0...main].
 - Show number of books in library ([#12])\
   The total number of books is now displayed at the top of the library. When searching or applying filters, you'll see
   both the number of matching books and the total. If books are grouped, the count for each group will also be shown.
+- Support Daily Notes via Periodic Notes plugin ([#18])\
+  Bookshelf now also supports Daily Notes via the Periodic Notes plugin.
 
 ### Fixed
 
@@ -193,4 +195,5 @@ For a full diff see [`0.2.2...0.3.0`][0.2.2...0.3.0].
 [#7]: https://github.com/weph/obsidian-bookshelf/issues/7
 [#8]: https://github.com/weph/obsidian-bookshelf/issues/8
 [#12]: https://github.com/weph/obsidian-bookshelf/issues/12
+[#18]: https://github.com/weph/obsidian-bookshelf/issues/18
 [@soundslikeinfo]: https://github.com/soundslikeinfo